+++ libethereum/BlockChain.h
@@ -178,7 +178,7 @@ public:
 
     LastBlockHashesFace const& lastBlockHashes() const { return *m_lastBlockHashes;  }
 
-    int chainID() const { return m_params.chainID; }
+    uint64_t chainID() const { return m_params.chainID; }
 
     /** Get the block blooms for a number of blocks. Thread-safe.
      * @returns the object pertaining to the blocks:
